requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
# Editable install with no version control (fakear==0.1)
-e /home/coton/Documents/Online/Projects/fakear
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.2.0
PyYAML==3.13
six==1.12.0
voluptuous==0.11.5
